--- OpenRA.Mods.Common/UtilityCommands/Extensions.cs
@@ -1,29 +1,0 @@
-#region Copyright & License Information
-/*
- * Copyright (c) The OpenRA Developers and Contributors
- * This file is part of OpenRA, which is free software. It is made
- * available to you under the terms of the GNU General Public License
- * as published by the Free Software Foundation, either version 3 of
- * the License, or (at your option) any later version. For more
- * information, see COPYING.
- */
-#endregion
-
-using System;
-using System.Collections.Generic;
-
-namespace OpenRA.Mods.Common.UtilityCommands
-{
-	public static class Extensions
-	{
-		public static IEnumerable<TSource> DistinctBy<TSource, TKey>(this IEnumerable<TSource> source, Func<TSource, TKey> keySelector)
-		{
-			var knownKeys = new HashSet<TKey>();
-			foreach (var element in source)
-			{
-				if (knownKeys.Add(keySelector(element)))
-					yield return element;
-			}
-		}
-	}
-}
